+++ nutsflow/__init__.py
@@ -11,7 +11,7 @@ from nutsflow.processor import (Take, Slice, Concat, Interleave, Zip, ZipWith,
                                 MapCol, MapMulti, MapPar, Prefetch,
                                 PrintProgress, Try)
 from nutsflow.function import (Identity, Square, NOP, Get, GetCols, Counter,
-                               Sleep, Format, Print)
+                               Sleep, Format, Print, PrintColType, PrintType)
 from nutsflow.sink import (Sort, Sum, Mean, MeanStd, Max, Min, ArgMax, ArgMin,
                            Reduce, Nth, Next, Consume, Count, Unzip, Head, Tail,
                            CountValues, Collect, Join, WriteCSV)